+++ packages/visualizations/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.32.1-beta.1](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations@0.32.1-beta.0...@opendatasoft/visualizations@0.32.1-beta.1) (2026-04-17)
+
+**Note:** Version bump only for package @opendatasoft/visualizations
+
+
+
+
+
+## [0.32.1-beta.0](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations@0.32.0...@opendatasoft/visualizations@0.32.1-beta.0) (2026-04-16)
+
+
+### Features
+
+* support action items and href onClick ([d794795](https://github.com/opendatasoft/ods-dataviz-sdk/commit/d79479599d892d005be8245f1f47a57e59fea8c2))
+
+
+
+
+
 # [0.32.0](https://github.com/opendatasoft/ods-dataviz-sdk/compare/@opendatasoft/visualizations@0.30.0...@opendatasoft/visualizations@0.32.0) (2026-04-08)